// src/shared/auth/auth.dto.ts
import { z } from 'zod'
import { ApiUserSchema } from './user.dto'

export const ApiRegisterPayloadSchema = z.object({
  user_name: z.string(),
  password: z.string(),
})

export const ApiAuthResponseSchema = z.object({
  user: ApiUserSchema,
  token: z.string(),
})

export const ApiLoginPayloadSchema = z.object({
  user_name: z.string(),
  password: z.string(),
  remember_me: z.boolean(),
})